materials/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "materials"
dynamic = ["version"]
description = ''
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = []
authors = [
  { name = "Leo", email = "leonardozandrade@gmail.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "qdrant-client",
  "python-dotenv",
  "sentence-transformers",
  "transformers",
  "llama-index-core",
  "llama-index-embeddings-huggingface",
  "datasets",
  "tqdm",
  "langchain-openai",
]

[project.urls]
Documentation = "https://github.com/Leo/materials#readme"
Issues = "https://github.com/Leo/materials/issues"
Source = "https://github.com/Leo/materials"

[tool.hatch.version]
path = "src/materials/__about__.py"

[tool.hatch.envs.types]
extra-dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/materials tests}"

[tool.coverage.run]
source_pkgs = ["materials", "tests"]
branch = true
parallel = true
omit = [
  "src/materials/__about__.py",
]

[tool.coverage.paths]
materials = ["src/materials", "*/materials/src/materials"]
tests = ["tests", "*/materials/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
